+++ windrecorder/record_wintitle.py
@@ -5,9 +5,12 @@ import re
 
 import pandas as pd
 import pygetwindow
+import streamlit as st
 
 from windrecorder import file_utils, utils
 from windrecorder.config import config
+from windrecorder.oneday import OneDay
+from windrecorder.utils import get_text as _t
 
 CSV_TEMPLATE_DF = pd.DataFrame(columns=["datetime", "window_title"])
 window_title_last_record = ""
@@ -87,3 +90,54 @@ def optimize_wintitle_name(text):
     text = re.sub("^\\(\\d+\\) ", "", text)  # 移除最开始的当前对话未读消息
 
     return text
+
+
+# 获取当天前台窗口标题统计情况
+def get_wintitle_stat_in_day(dt_in: datetime.datetime):
+    df = OneDay().search_day_data(dt_in, search_content="")
+    # 在生成前清洗数据：
+    # from windrecorder import record_wintitle
+    # df["win_title"] = df["win_title"].apply(record_wintitle.optimize_wintitle_name)
+    df.sort_values(by="videofile_time", ascending=True, inplace=True)
+    df = df.reset_index(drop=True)
+    stat = {}
+    for index, row in df.iterrows():
+        win_title_name = str(row["win_title"])
+        if win_title_name == "None" or win_title_name == "nan":
+            continue
+        if win_title_name not in stat:
+            stat[win_title_name] = 0
+        if index == df.index.max():
+            break
+        second_interval = df.loc[index + 1, "videofile_time"] - df.loc[index, "videofile_time"]
+        if second_interval > 100:  # 添加阈值，排除时间差值过大的 row，比如隔夜、锁屏期间的记录等
+            second_interval = 100
+        stat[win_title_name] += second_interval
+
+    # 清洗整理数据
+    stat = {key: val for key, val in stat.items() if val > 1}
+    df_show = pd.DataFrame(list(stat.items()), columns=["Page", "Screen Time"])
+    df_show.sort_values(by="Screen Time", ascending=False, inplace=True)
+    df_show = df_show.reset_index(drop=True)
+    df_show["Screen Time"] = df_show["Screen Time"].apply(utils.convert_seconds_to_hhmmss)
+
+    return df_show
+
+
+# ------------streamlit component
+# 窗口标题组件
+def component_wintitle_stat(day_date_input):
+    day_wintitle_df_statename_date = day_date_input.strftime("%Y-%m-%d")
+    day_wintitle_df_statename = f"wintitle_stat_{day_wintitle_df_statename_date}"
+    if day_wintitle_df_statename not in st.session_state:
+        st.session_state[day_wintitle_df_statename] = get_wintitle_stat_in_day(day_date_input)
+    if len(st.session_state[day_wintitle_df_statename]) > 0:
+        st.dataframe(
+            st.session_state[day_wintitle_df_statename],
+            column_config={"Page": st.column_config.TextColumn(_t("oneday_wt_text"), help=_t("oneday_wt_help"))},
+            height=650,
+            hide_index=True,
+            use_container_width=True,
+        )
+    else:
+        st.markdown(_t("oneday_ls_text_no_wintitle_stat"), unsafe_allow_html=True)
